Lazily load podman executable
When we introduced the ability to specify podman executable, we
inadvertently changed the behavior of the loading mechanism (commit
e120aa12fbce7fce2f483f01b24d57de61c67873).

Before that change, it was possible to initialize the podman driver even
if the podman executable was not present. In other words, it was OK to
install the podman driver without having podman installed on the system.

With the new mechanism in place, Molecule will fail to start when the
podman driver is installed and there is no podman executable available
on the system.

This commit restores the od lazy-loading behavior.