Allow to additionally configure arbitrary command-line options
It would be a significant maintenance burden to keep the set of configurable
attributes up-to-date. So, while providing an easy way to set essential
arguments with dedicated attributes, with the proposed change any available
option implemented by Podman can be utilized.

This change is inspired by Ansible's unarchive module [1], its attribute
'extra_opts' forwards additional options to the underlying CLI,
regardless of implementation of the CLI.

[1] https://docs.ansible.com/ansible/latest/collections/ansible/builtin/unarchive_module.html#parameter-extra_opts