Skip podman network creation named with 'bridge'
As seen in https://docs.podman.io/en/latest/markdown/podman-run.1.html#network-mode-net "bridge" is a special name used to create a network stack on the default bridge. 
When a "bridge" podman network is created, it won't be used when  `podman run --network=bridge ...` is executed.
The same happends with 'none', 'host', 'ns', 'private', 'slirp4netns' it won't be used or it will fail as incomplete argument.